:recycle: extract line items class + fix taxes iteration syntax
